backend/requirements.txt
# ====================================
# SpitKorean Backend Dependencies
# ====================================

# 웹 프레임워크
quart==0.18.4
hypercorn==0.14.4
quart-cors==0.7.0  # CORS 처리를 위해 추가
blinker>=1.9.0  # Flask 호환성을 위해 최신 버전 명시

# 비동기 데이터베이스
motor==3.3.1  # MongoDB 비동기 드라이버 (main.py에서 사용)
pymongo==4.5.0

# 캐시 및 작업 큐
celery==5.3.1
redis[hiredis]==5.0.1  # EventBus, CacheManager, UsageLimiter에서 사용
aioredis>=2.0.1,<3.0.0  # 비동기 Redis 클라이언트 (추가 호환성)

# AI/ML 서비스
openai==1.3.7  # GPTService, WhisperService에서 사용
google-cloud-texttospeech==2.14.1  # TTSService에서 사용
google-cloud-speech==2.21.0  # 음성 인식용
google-auth==2.23.0  # Google Cloud 인증

# 데이터 처리 및 검증
pydantic==1.10.13  # config.py에서 BaseSettings 사용 (Rust 없이 설치 가능)
pymongo[srv]==4.5.0  # MongoDB 연결 (srv 지원 포함)
jsonschema==4.20.0  # JSON 스키마 검증

# 인증 및 보안
pyjwt==2.8.0  # AuthManager에서 JWT 토큰 처리
bcrypt==4.1.2  # auth.py에서 비밀번호 해싱

# HTTP 클라이언트 및 네트워킹
httpx==0.24.1  # 비동기 HTTP 요청
aiohttp>=3.8.4,<4.0.0  # 추가 HTTP 클라이언트 지원
requests==2.31.0  # 동기 HTTP 요청 (일부 서비스에서 필요)

# 오디오/이미지 처리
pydub==0.25.1  # WhisperService에서 오디오 처리
pillow==11.2.1  # 이미지 처리
numpy==2.2.6  # 수치 계산

# 유틸리티
python-dotenv==1.0.0  # 환경 변수 로드 (main.py에서 사용)
tenacity==8.2.3  # 재시도 로직 (@retry 데코레이터)
python-multipart==0.0.6  # 멀티파트 폼 데이터 처리
structlog==23.2.0  # 구조화된 로깅
aiofiles==23.2.1  # 비동기 파일 처리
werkzeug>=3.1.3  # 유틸리티 함수

# 이메일 및 알림
email-validator==2.1.0  # 이메일 유효성 검증 (validators.py용)
# smtplib는 Python 내장 모듈이므로 별도 설치 불필요

# 스케줄링 및 백그라운드 작업
# celery[redis]는 이미 위에서 설치되므로 중복 제거

# 데이터 유효성 검증 및 포맷팅
validators==0.22.0  # 추가 검증 유틸리티
python-dateutil==2.8.2  # 날짜 처리

# 개발 및 테스트 도구
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1
pytest-cov==4.1.0
black==23.3.0
isort==5.12.0
mypy==1.3.0

# 타입 힌트 지원
types-requests==2.31.0
types-redis==4.6.0

# 프로덕션 서버
gunicorn==21.2.0  # WSGI 서버 (선택적)
uvicorn[standard]==0.23.2  # ASGI 서버 (Quart용 대안)

# 모니터링 및 로깅
sentry-sdk[quart]==1.38.0  # 에러 모니터링 (선택적)

# 빌드 도구 (Python 3.13 호환성)
setuptools>=65.0.0
wheel>=0.38.0

# ====================================
# 주요 변경사항 및 추가된 패키지:
# ====================================
# 1. quart-cors: CORS 처리 (main.py의 CORS 헤더 설정용)
# 2. pydantic: config.py에서 BaseSettings 사용
# 3. google-auth: Google Cloud 서비스 인증
# 4. bson: ObjectId 처리 (모든 라우트에서 사용)
# 5. email-validator: 이메일 검증 (validators.py용)
# 6. smtplib-async: 비동기 이메일 전송 (notification_tasks.py용)
# 7. validators: 추가 검증 유틸리티
# 8. python-dateutil: 날짜 처리
# 9. requests: 일부 동기 API 호출용
# 10. uvicorn: Quart ASGI 서버 대안
# 11. sentry-sdk: 프로덕션 에러 모니터링 (선택적)
# 12. types-* 패키지들: mypy 타입 체킹 지원